refactor: move query use cases into application layer
